Isolate TAPI tests validating resource-based test execution
