+++ platforms/ide/tooling-api/build.gradle.kts
@@ -79,7 +79,6 @@ dependencies {
     testFixturesImplementation(testFixtures(projects.enterpriseLogging))
     testFixturesImplementation(libs.commonsIo)
     testFixturesImplementation(libs.slf4jApi)
-    testFixturesImplementation(testFixtures(projects.testingJvm))
 
     integTestImplementation(projects.jvmServices)
     integTestImplementation(projects.persistentCache)
@@ -92,7 +91,6 @@ dependencies {
     crossVersionTestImplementation(testFixtures(projects.problemsApi))
     crossVersionTestImplementation(libs.jettyWebApp)
     crossVersionTestImplementation(libs.commonsIo)
-    crossVersionTestImplementation(testFixtures(projects.testingJvm))
     crossVersionTestRuntimeOnly(libs.cglib) {
         because("BuildFinishedCrossVersionSpec classpath inference requires cglib enhancer")
     }
