--- platforms/jvm/testing-jvm/src/testFixtures/resources/testengines/custom-test-source/test-engine-build/src/main/resources/META-INF/services/org.junit.platform.engine.TestEngine
@@ -1,1 +1,0 @@
-org.gradle.testing.testengine.engine.CustomTestSourceDeclaringTestEngine
